fix: move parsing function to common utils
